added: destroy answers with ajax

+++ app/controllers/answers_controller.rb
@@ -10,17 +10,12 @@ class AnswersController < ApplicationController
   end
 
   def update
-    answer.update(answer_params)
+    answer.update(answer_params) if current_user.author?(answer)
     @question = answer.question
   end
 
   def destroy
-    if current_user&.author?(answer)
-      answer.destroy
-      redirect_to question_path(answer.question), notice: 'Answer successfully removed.'
-    else
-      redirect_to question_path(answer.question), notice: 'Answer can be deleted only by the author.'
-    end
+    answer.destroy if current_user.author?(answer)
   end
 
   private
